# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateAccountRequest:
    properties:
      document_number:
        example: "1234567890"
        type: string
    type: object
  dto.CreateAccountResponse:
    properties:
      id:
        example: 1
        type: integer
    type: object
  dto.CreateTransactionRequest:
    properties:
      account_id:
        example: 1
        type: integer
      amount:
        example: 100
        type: number
      operation_type_id:
        example: 4
        type: integer
    type: object
  dto.CreateTransactionResponse:
    properties:
      id:
        example: 1
        type: integer
    type: object
  dto.GetAccountResponse:
    properties:
      account_id:
        example: 1
        type: integer
      document_number:
        example: "1234567890"
        type: string
    type: object
  response.ErrorResponse:
    properties:
      description:
        type: string
      error:
        type: string
      status_code:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for account and transaction management
  title: Transaction Flow API
  version: "1.0"
paths:
  /accounts:
    post:
      consumes:
      - application/json
      description: Creates a new account with a document number
      parameters:
      - description: Account creation request
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Account Created
          schema:
            $ref: '#/definitions/dto.CreateAccountResponse'
        "400":
          description: Invalid Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Account Already Exists
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Validation Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create an account
      tags:
      - Accounts
  /accounts/{id}:
    get:
      consumes:
      - application/json
      description: Fetches account details by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Account Details
          schema:
            $ref: '#/definitions/dto.GetAccountResponse'
        "404":
          description: Account Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Retrieve an account
      tags:
      - Accounts
  /transactions:
    post:
      consumes:
      - application/json
      description: Registers a new financial transaction
      parameters:
      - description: Transaction Request
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Transaction Created
          schema:
            $ref: '#/definitions/dto.CreateTransactionResponse'
        "400":
          description: Invalid Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Validation Failed
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a transaction
      tags:
      - Transactions
schemes:
- http
swagger: "2.0"